- Add basic elements   - Fragments   - Navigation component   - Organize structure project - Fix libraries bugs - Fix AndroidManifest - Add module
