+++ .idea/misc.xml
@@ -3,6 +3,10 @@
   <component name="DesignSurface">
     <option name="filePathToZoomLevelMap">
       <map>
+        <entry key="..\:/Users/LidiaPC/Desktop/apps/PruebaMercadona/app/src/main/res/layout/fragment_detail.xml" value="0.10989583333333333" />
+        <entry key="..\:/Users/LidiaPC/Desktop/apps/PruebaMercadona/app/src/main/res/layout/fragment_list.xml" value="0.11927083333333334" />
+        <entry key="..\:/Users/LidiaPC/Desktop/apps/PruebaMercadona/app/src/main/res/layout/fragment_splash.xml" value="0.11927083333333334" />
+        <entry key="..\:/Users/LidiaPC/Desktop/apps/PruebaMercadona/app/src/main/res/layout/nav_host_activity.xml" value="0.17404898384575299" />
         <entry key="..\:/Users/LidiaPC/Desktop/apps/PruebaMercadona/app/src/main/res/mipmap-anydpi-v26/ic_launcher.xml" value="0.1095" />
         <entry key="..\:/Users/LidiaPC/Desktop/apps/PruebaMercadona/app/src/main/res/mipmap-anydpi-v26/ic_launcher_round.xml" value="0.1095" />
       </map>
