+++ app/src/main/AndroidManifest.xml
@@ -4,13 +4,14 @@
 
     <application
         android:allowBackup="true"
+        android:name=".di.MyApplication"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.PruebaMercadona">
         <activity
-            android:name=".NavHostActivity"
+            android:name=".ui.activity.NavHostActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
