+++ app/src/main/java/com/lidorttol/pruebamercadona/di/AppModule.kt
@@ -1,0 +1,11 @@
+package com.lidorttol.pruebamercadona.di
+
+import dagger.Module
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+
+@Module
+@InstallIn(SingletonComponent::class)
+object AppModule {
+
+}
