+++ app/src/main/java/com/lidorttol/pruebamercadona/di/MyApplication.kt
@@ -1,0 +1,7 @@
+package com.lidorttol.pruebamercadona.di
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+
+@HiltAndroidApp
+class MyApplication : Application()
