+++ app/src/main/java/com/lidorttol/pruebamercadona/ui/activity/NavHostActivity.kt
@@ -1,11 +1,16 @@
-package com.lidorttol.pruebamercadona
+package com.lidorttol.pruebamercadona.ui.activity
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import com.lidorttol.pruebamercadona.R
+import dagger.hilt.android.AndroidEntryPoint
 
+@AndroidEntryPoint
 class NavHostActivity : AppCompatActivity() {
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.nav_host_activity)
     }
+
 }
