+++ app/src/main/java/com/lidorttol/pruebamercadona/ui/fragment/DetailFragment.kt
@@ -1,0 +1,19 @@
+package com.lidorttol.pruebamercadona.ui.fragment
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.lidorttol.pruebamercadona.R
+
+class DetailFragment : Fragment() {
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        return inflater.inflate(R.layout.fragment_detail, container, false)
+    }
+
+}
