+++ app/src/main/java/com/lidorttol/pruebamercadona/ui/viewmodel/MainViewModel.kt
@@ -1,0 +1,11 @@
+package com.lidorttol.pruebamercadona.ui.viewmodel
+
+import android.app.Application
+import androidx.lifecycle.AndroidViewModel
+import dagger.hilt.android.lifecycle.HiltViewModel
+import javax.inject.Inject
+
+@HiltViewModel
+class MainViewModel @Inject constructor(
+    application: Application
+) : AndroidViewModel(application)
